Attempt to detect invalid address and don't send to it
In the case where a connected app crashes, then they won't have disconnected
properly. Because of this, we might still be attempting to send updates to it,
which causes the underlying OS to complain about the socket being closed (and
sometimes causes LiSP to crash).

It might be preferable to detect socket disconnections and do something about
them then, and if we were using `liblo` directly that might be achievable, but
that's not (I think) possible down to the level of abstraction.